Feature- View top 5 actors.
Feature:As a user I want to be able to view top 5 actors that are part of films I have in the store
Got rid of test.js since I no longer need it!
Next added and connected the frontend to the backend for this endpoint, was really fast and then i added the genres and stuff which took a bit longer but was fine in the end.