Feature 4/7 See top 5 actors Details
Feature: As a user I want to be able to view the actor’s details and view their top 5 rented films

I create another page for actors and it routes correctly and I connected it easily since the data was preloaded. However, doing this made me realize that I do need to change how i configure the data, and that I should definitely set up an API endpoint in the backend to get details specifically. I also added some extra features that weren't necessary but I thought would be nice. Next commit will be adding that endpoint and properly configuring it.