Feature 5/7
Feature: As a user I want to be able to view details of the film

You can view the details of a film when you click on them! This will be applicable in any scenerio so it ends up working and I also added back tracking functionality because it got annoying have to click my way back to where I was. I need to add the top 5 thing when you click on the top 5 films, I'll do that next.