Feature 6/7 - Movie search by Title,Genre,Actor
Feature: As a user I want to be able to search a film by name of film, name of an actor, or genre of the film

This implementation was interesting since it was unique in the sense of having to understand the whole timeout/polling sequence but besides that was pretty straight forward. I also just realized but I don't really have information about rentals and how many are left so I will have to add that to the moviesScreen.