Feature 7/7 - rent to customer
Feature: As a user I want to be able to rent a film out to a customer

This was problably the most annoying considering I had to create everything in a way I wanted.
I successfully implemented the ability to rent to a customer using the Customer ID, while I could use the name, people can technically have the same name so this is the best choice.