Feature 8/14 - View all Customer
Use case: As a user I want to view a list of all customers (Pref. using pagination)

This was pretty straightforward, I decided to use a grid rather than a flexbox model because I wanted to see the difference, and decided to stick with the grid. Beyond that, setting up the functions was definitely interesting to debug.