Feature 9/14 Search customers:
Feature: As a user I want the ability to filter/search customers by their customer id, first name or last name.

I changed the frontend design like 3 times because it was so crowded and had a lot of div's going on so I moved things into separate components so I could actually read.
Also made and moved css to fit the new components.