Feature 12/14 Delete Customer
Feature: As a user I want to be able to delete a customer if they no longer wish to patron at store

Overall, as I'm nearing the end of this project, I'm getting a bit more lazy with trying to modularize stuff, but that's okay because there's only 2 features left. Anyway, the frontend once again was the most annoying thing to implement and actually handle, but I've gotten the hang of speeding through it well.