Feature 14/14 Indicate Customer has return a film
Feature: As a user I want to be able to indicate that a customer has returned a rented movie

This feature in the frontend once again pretty straightforward considering all the pieces were right there to finish it, simply go to the history and you'd be able to return the respective film you need. I also got rid of the about tab on the top because it was staight up useless since I'm tired and don't feel like implementing that.